requirements.txt
aiohttp>=3.6.2
aiosqlite>=0.15.0
discord.py==1.6.0
